Move some tests around and add test for sel
