+++ tests/cassettes/test_rasterio_backend/test_rasterio_statistics.yaml
