+++ tests/cassettes/test_xarray_backend/test_timeseries_statistics.yaml
