+++ tests/cassettes/test_xarray_backend/test_timeseries_statistics_image_size_limit.yaml
